Add new sgd api.
To allow for this API we need to have the params already being traced.
Thus this change always automatically enables tracing for every
initialized param. Furthermore, there needs to be a tape on the stack
waiting, if any top-level params get traced.

This adds a function gradParams2 which should be renamed at some point,
but since this is an experiemental API, I'll stick with it for now.